+++ src/Builder.fs
@@ -2,19 +2,61 @@ namespace RunHelpers
 
 [<AutoOpen>]
 module Builder =
-    type JobBuilder(combine) =
-        member __.Combine(process1, process2) = combine process1 process2
+    type Delayed<'a> = unit -> 'a
+    type DelayedResult = Delayed<JobResult>
 
+    [<AbstractClass>]
+    type BaseJobBuilder() =
+        member __.Zero() = Ok
+
+    type SequentialJobBuilder() =
+        inherit BaseJobBuilder()
+
+        member __.Combine(jobResult: JobResult, job2: DelayedResult) =
+            match jobResult with
+            | Ok -> job2 ()
+            | Error _ -> jobResult
+
+        member __.Delay f : Delayed<'a> = f
+
+        member this.For(collection: 'a seq, evaluation: 'a -> JobResult) =
+            let delayedEvaluate value = (fun () -> evaluation value)
+
+            collection
+            |> Seq.fold (fun beforeResult value -> this.Combine(beforeResult, delayedEvaluate value)) Ok
+
+        member __.Run(f: DelayedResult) : Job = Job.create f
+
+        member __.Yield x : JobResult = x
+        member __.Yield(job: Job) : JobResult = job |> Async.RunSynchronously
+
+    type ParallelJobBuilder() =
+        inherit BaseJobBuilder()
+
+        member __.Combine(job1: Job, job2: Job) = [ job2; job1 ]
+        member __.Combine(jobList: Job list, job: Job) = job :: jobList
         member __.Delay f = f ()
 
-        member __.For(lst, f) =
-            lst
-            |> Seq.fold (fun f1 el -> combine f1 (f el)) Job.ok
+        member this.For(collection: 'a seq, evaluation: 'a -> Job) =
+            collection
+            |> Seq.fold (fun (jobList: Job list) value -> this.Combine(jobList, evaluation value)) []
 
-        member __.Run f = f
+        member __.Run(jobList: Job seq) : Job =
+            async {
+                let! resultList = Async.Parallel jobList
 
-        member __.Yield x : Job = x
-        member __.Zero() : Job = Job.ok
+                return
+                    Array.fold
+                        (fun combinedResult result ->
+                            match combinedResult, result with
+                            | Ok, result -> result
+                            | combinedResult, Ok -> combinedResult
+                            | Error msgList1, Error msgList2 -> Error(List.append msgList1 msgList2))
+                        Ok
+                        resultList
+            }
 
-    let job = JobBuilder(Job.combineSequential)
-    let parallelJob = JobBuilder(Job.combineParallel Constant.errorExitCode)
+        member __.Yield job : Job = job
+
+    let job = SequentialJobBuilder()
+    let parallelJob = ParallelJobBuilder()