+++ src/Constant.fs
@@ -1,5 +1,0 @@
-namespace RunHelpers
-
-[<RequireQualifiedAccess>]
-module Constant =
-    let errorExitCode = 10