+++ src/Types.fs
@@ -2,15 +2,16 @@
 
 type JobResult =
     | Ok
-    | Error of exitCode: int * msg: string list
+    | Error of msg: string list
 
 type Job = Async<JobResult>
 
 module Job =
+    let errorExitCode = 1
+
     let ok = async { return Ok }
 
-    let error exitCode msgList =
-        async { return Error(exitCode, msgList) }
+    let error msgList = async { return Error(msgList) }
 
     let allowFailure (job: Job) : Job =
         async {
@@ -18,36 +19,7 @@ module Job =
             return Ok
         }
 
-    /// Combines jobs sequentially. If the first job suceeds, the second is run,
-    /// otherwise it returns a Error JobResult
-    let combineSequential job1 job2 : Job =
-        async {
-            let! result1 = job1
-
-            match result1 with
-            | Ok -> return! job2
-            | Error (code, msg) -> return Error(code, msg)
-        }
-
-    let combineParallelMany multiErrorExitCode jobList =
-        async {
-            let! results = Async.Parallel jobList
-
-            return
-                Array.fold
-                    (fun totalResult jobResult ->
-                        match totalResult, jobResult with
-                        | Ok, Ok -> Ok
-                        | Error (exitCode, msgList), Ok
-                        | Ok, Error (exitCode, msgList) -> Error(exitCode, msgList)
-                        | Error (_, msgList1), Error (_, msgList2) ->
-                            Error(multiErrorExitCode, List.append msgList1 msgList2))
-                    Ok
-                    results
-        }
-
-    let combineParallel multiErrorExitCode job1 job2 : Job =
-        combineParallelMany multiErrorExitCode [ job1; job2 ]
+    let create f : Job = async { return f () }
 
     /// Prints error messages and returns exit code
     let execute job =
@@ -55,19 +27,21 @@ module Job =
 
         match result with
         | Ok -> 0
-        | Error (x, msg) ->
+        | Error msg ->
             // We print every msg the error has
             List.iter (printfn "%s") msg
-            x
+            errorExitCode
 
     open Fake.Core
 
     /// Creates a job from a FAKE CreateProcess instance
-    let fromCreateProcess errorCode (proc: CreateProcess<ProcessResult<unit>>) : Job =
+    let fromCreateProcess (proc: CreateProcess<ProcessResult<unit>>) : Job =
         async {
-            printfn $"> %s{proc.CommandLine}"
+            // TODO: move to allow customization like [parallel]-Prefix
+            // explicit \n to improve parallel behaviour
+            printfn "> %s" proc.CommandLine
 
             match (Proc.run proc).ExitCode with
             | 0 -> return Ok
-            | _ -> return Error(errorCode, [])
+            | _ -> return Error([ "Execution failed!" ])
         }