+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Replaced `Template` module with RequiredQualifiedAccess by `Templates` namespace
 - Restructure structure to enable parallelization feature
+- Remove exitcode from error type
 
 ## [0.4.1] - 2021-11-16
 ### Fixed